Updated list component markup
With this commit I added global elements to the list component. The same way we prepared the default markup and the CSS for content cards (e.g. product card), we should have a default markup and styles for the list items. It should be generic and usable even without modifications but allow flexibility to change it to match the requirements by 80% of themes. If a theme needs more specific design then it should have a custom implementation.