[Headline] Added pagination to the archive page
refs https://github.com/TryGhost/Team/issues/2669

The archive page used to display all posts at once using get helper. For sites with large volumes of content this can cause the site to timeout and render a blank page. This fixes the problem by fetching the posts of the first page manually and adding pagination to the rest of the content.